# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraShield AI - Women Safety Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatInput {
  /** User message */
  message: string;
  /** Username for memory context */
  username: string;
}

/**
 * Safety risk classification
 */
export type ChatResponseRiskLevel = typeof ChatResponseRiskLevel[keyof typeof ChatResponseRiskLevel];


export const ChatResponseRiskLevel = {
  SAFE: 'SAFE',
  MODERATE: 'MODERATE',
  HIGH_RISK: 'HIGH RISK',
} as const;

/**
 * User memory context
 */
export type ChatResponseMemory = { [key: string]: unknown };

export interface ChatResponse {
  /** AI response text */
  reply: string;
  /** Whether emergency was detected */
  emergency: boolean;
  /** Safety risk classification */
  risk_level: ChatResponseRiskLevel;
  /** AI model used */
  model: string;
  /** User memory context */
  memory?: ChatResponseMemory;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
  timestamp: string;
  /** @nullable */
  risk_level?: string | null;
  /** @nullable */
  emergency?: boolean | null;
}

export interface ChatHistory {
  username: string;
  messages: ChatMessage[];
}

export interface SafetyStatus {
  username: string;
  risk_level: string;
  emergency_count: number;
  unsafe_locations?: string[];
  /** @nullable */
  last_active?: string | null;
}

export interface SosInput {
  username: string;
  /** @nullable */
  location?: string | null;
}

export interface EmergencyContact {
  name: string;
  phone: string;
  relation?: string;
}

export interface SosResponse {
  success: boolean;
  message: string;
  reply: string;
  contacts?: EmergencyContact[];
}

export interface ContactsResponse {
  username: string;
  contacts: EmergencyContact[];
}

export interface AddContactBody {
  username: string;
  name: string;
  phone: string;
  relation?: string;
}

export type GetChatHistoryParams = {
username: string;
};

export type GetSafetyStatusParams = {
username: string;
};

export type GetContactsParams = {
username: string;
};

export type DeleteContactParams = {
username: string;
};