# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraShield AI - Women Safety Assistant API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Send a user message and receive AI response with safety analysis
 * @summary Send a message to AuraShield AI
 */
export const SendMessageBody = zod.object({
  "message": zod.string().describe('User message'),
  "username": zod.string().describe('Username for memory context')
})

export const SendMessageResponse = zod.object({
  "reply": zod.string().describe('AI response text'),
  "emergency": zod.boolean().describe('Whether emergency was detected'),
  "risk_level": zod.enum(['SAFE', 'MODERATE', 'HIGH RISK']).describe('Safety risk classification'),
  "model": zod.string().describe('AI model used'),
  "memory": zod.record(zod.string(), zod.unknown()).optional().describe('User memory context')
})


/**
 * @summary Get chat history for a user
 */
export const GetChatHistoryQueryParams = zod.object({
  "username": zod.coerce.string()
})

export const GetChatHistoryResponse = zod.object({
  "username": zod.string(),
  "messages": zod.array(zod.object({
  "role": zod.enum(['user', 'assistant']),
  "content": zod.string(),
  "timestamp": zod.string(),
  "risk_level": zod.string().nullish(),
  "emergency": zod.boolean().nullish()
}))
})


/**
 * @summary Get current safety status and memory for a user
 */
export const GetSafetyStatusQueryParams = zod.object({
  "username": zod.coerce.string()
})

export const GetSafetyStatusResponse = zod.object({
  "username": zod.string(),
  "risk_level": zod.string(),
  "emergency_count": zod.number(),
  "unsafe_locations": zod.array(zod.string()).optional(),
  "last_active": zod.string().nullish()
})


/**
 * @summary Trigger SOS emergency alert
 */
export const TriggerSosBody = zod.object({
  "username": zod.string(),
  "location": zod.string().nullish()
})

export const TriggerSosResponse = zod.object({
  "success": zod.boolean(),
  "message": zod.string(),
  "reply": zod.string(),
  "contacts": zod.array(zod.object({
  "name": zod.string(),
  "phone": zod.string(),
  "relation": zod.string().optional()
})).optional()
})


/**
 * @summary Get emergency contacts for a user
 */
export const GetContactsQueryParams = zod.object({
  "username": zod.coerce.string()
})

export const GetContactsResponse = zod.object({
  "username": zod.string(),
  "contacts": zod.array(zod.object({
  "name": zod.string(),
  "phone": zod.string(),
  "relation": zod.string().optional()
}))
})


/**
 * @summary Add an emergency contact
 */
export const AddContactBody = zod.object({
  "username": zod.string(),
  "name": zod.string(),
  "phone": zod.string(),
  "relation": zod.string().optional()
})

export const AddContactResponse = zod.object({
  "username": zod.string(),
  "contacts": zod.array(zod.object({
  "name": zod.string(),
  "phone": zod.string(),
  "relation": zod.string().optional()
}))
})


/**
 * @summary Delete an emergency contact by index
 */
export const DeleteContactParams = zod.object({
  "index": zod.coerce.number()
})

export const DeleteContactQueryParams = zod.object({
  "username": zod.coerce.string()
})

export const DeleteContactResponse = zod.object({
  "username": zod.string(),
  "contacts": zod.array(zod.object({
  "name": zod.string(),
  "phone": zod.string(),
  "relation": zod.string().optional()
}))
})